Despeckling now applies to closed linestrings too to help remove small closed pressure curve (BRAINSTORM-1497)
Note that one should not apply despecling simultaneously to isobands and isolines, since any isoband
touching the boundary of the area is closed, but the isoline is not. The isoband may thus be removed
while the isoline is not. Despeckling isobands only or isolines only should work fine for most
despeckling purposes.