Added winding order handling
OGR::normalizeWindingOrder will change the geometry so that all
polygon exteriors are clockwise and interiors counter-clockwise.

OGR::renormalizeWindingOrder will take a geometry that was normalized
before a projection and fix the geometry so that any exterior that has
reversed its winding order will become a hole instead.

Sample use case:

  - polar stereographic (north pole) projection with Antarctic polygon data
  - the Antarctic polygon circles the south pole which recedes to infinity
    in the projection
  - the projected boundary of the Antarctic coastline circles all other continental
    shapes and will fill the entire map unless fixed

The fix is to take the envelope of the projected Antarctic, and add a new exterior
surrounding the projected Antarctic, which now becomes a hole. The exterior is
made by calculating the envelope of the projected Antarctic, and adding a box
around it triple the size of the envelope.