Fixed thread safety of spatial references
Setting the axis mapping strategy in the SpatialReference constructor is not
thread safe, since OGRSpatialReferenceFactory caches the objects and another
thread may access the cache immediately. The axis mapping strategy is now
set in the factory before the spatial reference is cached, so no other thread
has access to the spatial reference.

Setting the axis mapping strategy is not trivial, and leads to calling
proj_destroy, which makes the operation unsafe.